+++ src/Accordion/Accordion.wrapper.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { UUID } from '../ItemContainer/ItemContainer';
+import { UUID } from '../ItemContext/ItemContext';
 
 import {
     AccordionContainer,
