+++ src/AccordionItemHeading/AccordionItemHeading.spec.tsx
@@ -2,7 +2,7 @@ import { mount, ReactWrapper } from 'enzyme';
 import * as React from 'react';
 import { Provider as AccordionProvider } from '../AccordionContext/AccordionContext';
 import { Item } from '../AccordionStore/AccordionStore';
-import { Provider as ItemProvider } from '../ItemContainer/ItemContainer';
+import { Provider as ItemProvider } from '../ItemContext/ItemContext';
 import { default as AccordionItemHeading } from './AccordionItemHeading.wrapper';
 
 describe('AccordionItemHeading', () => {
