+++ src/AccordionItemPanel/AccordionItemPanel.wrapper.tsx
@@ -6,8 +6,8 @@ import {
 import { Item } from '../AccordionStore/AccordionStore';
 import {
     Consumer as ItemConsumer,
-    ItemContainer,
-} from '../ItemContainer/ItemContainer';
+    ItemContext,
+} from '../ItemContext/ItemContext';
 import AccordionItemPanel from './AccordionItemPanel';
 
 type AccordionItemPanelWrapperProps = React.HTMLAttributes<HTMLDivElement> & {
@@ -24,7 +24,7 @@ export default class AccordionItemPanelWrapper extends React.Component<
 
     renderChildren = (
         accordionStore: AccordionContainer,
-        itemStore: ItemContainer,
+        itemStore: ItemContext,
     ): JSX.Element => {
         const { uuid } = itemStore;
         const { items, allowMultipleExpanded } = accordionStore;
@@ -46,7 +46,7 @@ export default class AccordionItemPanelWrapper extends React.Component<
             <AccordionConsumer>
                 {(accordionStore: AccordionContainer): JSX.Element => (
                     <ItemConsumer>
-                        {(itemStore: ItemContainer): JSX.Element =>
+                        {(itemStore: ItemContext): JSX.Element =>
                             this.renderChildren(accordionStore, itemStore)
                         }
                     </ItemConsumer>
